+++ Musteriİslemleri.cs
@@ -19,12 +19,6 @@ namespace RentACar2023
             InitializeComponent();
             resetle();
         }
-        private void geriBTN_Click(object sender, EventArgs e)
-        {
-            GirisSayfasi girisSayfa = new GirisSayfasi();
-            girisSayfa.Show();
-            this.Hide();
-        }
         private void MusteriKayit_FormClosed(object sender, FormClosedEventArgs e)
         {
             Application.Exit();
@@ -33,72 +27,89 @@ namespace RentACar2023
         {
             if (sifreDegisTC.Text == "" || sifreDegisSifre.Text == "")
             {
-                MessageBox.Show("Lütfen boş alan bırakmayınız.");
+                MessageBox.Show("Lütfen boş alan bırakmayınız.", "UYARI", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
             else if (sifreDegisTC.Text.Length != 11)
             {
-                MessageBox.Show("Lütfen TC kimlik numaranızı eksiksiz bir şekilde giriniz.");
+                MessageBox.Show("Lütfen TC kimlik numaranızı eksiksiz bir şekilde giriniz.", "UYARI", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
             else
             {
-                string myConnectionString = "server=7xz.h.filess.io;database=rentacar_wastesugar;uid=rentacar_wastesugar;pwd=d150c35368dc92fa3cc2c09bde449b384fb6b4c3;port=3307;";
-                MySqlConnection cnn = new MySqlConnection(myConnectionString);
-                cnn.Open();
-                MySqlCommand sorgu = new MySqlCommand("SELECT * FROM users WHERE TC= '" + sifreDegisTC.Text + "'", cnn);
-                MySqlDataReader tara = sorgu.ExecuteReader();
-                if (tara.Read())
+                DialogResult result = MessageBox.Show("Şifrenizi değiştirmek istediğinize emin misiniz?", "Şifre Değiştirme Onayı", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+
+                if (result == DialogResult.Yes)
                 {
-                    string hashedPassword = BCrypt.Net.BCrypt.HashPassword(sifreDegisSifre.Text);
-                    cnn.Close();
+                    string myConnectionString = "server=7xz.h.filess.io;database=rentacar_wastesugar;uid=rentacar_wastesugar;pwd=d150c35368dc92fa3cc2c09bde449b384fb6b4c3;port=3307;";
+                    MySqlConnection cnn = new MySqlConnection(myConnectionString);
                     cnn.Open();
-                    MySqlCommand sorgu1 = new MySqlCommand("UPDATE users SET password ='" + hashedPassword + "' WHERE TC ='" + sifreDegisTC.Text + "' ", cnn);
-                    sorgu1.ExecuteNonQuery();
-                    MessageBox.Show("Şifreniz başarıyla değiştirildi");
-                    cnn.Close();
+                    MySqlCommand sorgu = new MySqlCommand("SELECT * FROM users WHERE TC= '" + sifreDegisTC.Text + "'", cnn);
+                    MySqlDataReader tara = sorgu.ExecuteReader();
+                    if (tara.Read())
+                    {
+                        string hashedPassword = BCrypt.Net.BCrypt.HashPassword(sifreDegisSifre.Text);
+                        cnn.Close();
+                        cnn.Open();
+                        MySqlCommand sorgu1 = new MySqlCommand("UPDATE users SET password ='" + hashedPassword + "' WHERE TC ='" + sifreDegisTC.Text + "' ", cnn);
+                        sorgu1.ExecuteNonQuery();
+                        MessageBox.Show("Şifreniz başarıyla değiştirildi", "Başarılı", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        cnn.Close();
+                    }
+                    else
+                    {
+                        MessageBox.Show("Böyle bir müşteri bulunamadı", "UYARI", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        cnn.Close();
+                    }
                 }
-                else
+                else if (result == DialogResult.No)
                 {
-                    MessageBox.Show("Böyle bir kullanıcı bulunamadı.");
-                    cnn.Close();
                 }
             }
         }
+
 
         private void musteriOlusturBTN_Click(object sender, EventArgs e)
         {
             if (molusturTC.Text == "" || molusturSifre.Text == "" || molusturAd.Text == "" || molusturSoyad.Text == "" || molusturTelNo.Text == "")
             {
-                MessageBox.Show("Lütfen boş alanları doldurup tekrar deneyiniz.");
+                MessageBox.Show("Lütfen boş alanları doldurup tekrar deneyiniz.", "UYARI", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
             else if (molusturTC.Text.Length != 11)
             {
-                MessageBox.Show("Lütfen TC kimlik numaranızı eksiksiz bir şekilde giriniz.");
+                MessageBox.Show("Lütfen TC kimlik numaranızı eksiksiz bir şekilde giriniz.", "UYARI", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
             else if (molusturTelNo.Text.Length != 11)
             {
-                MessageBox.Show("Lütfen telefon numaranızı eksiksiz bir şekilde giriniz.");
+                MessageBox.Show("Lütfen telefon numaranızı eksiksiz bir şekilde giriniz.", "UYARI", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
             else
             {
-                string myConnectionString = "server=7xz.h.filess.io;database=rentacar_wastesugar;uid=rentacar_wastesugar;pwd=d150c35368dc92fa3cc2c09bde449b384fb6b4c3;port=3307;";
-                MySqlConnection cnn = new MySqlConnection(myConnectionString);
-                cnn.Open();
-                MySqlCommand sorgu = new MySqlCommand("SELECT * FROM users WHERE TC= '" + molusturTC.Text + "'", cnn);
-                MySqlDataReader tara = sorgu.ExecuteReader();
-                if (tara.Read())
+                DialogResult result = MessageBox.Show("Müşteri eklemek istediğinize emin misiniz?", "Müşteri Ekleme Onayı", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+
+                if (result == DialogResult.Yes)
                 {
-                    MessageBox.Show("Böyle bir müşteri zaten mevcut, şifresini unuttuysanız yan taraftan şifresini değiştirebilirsiniz veya müşteriyi kaldırabilirsiniz.");
-                    cnn.Close();
-                }
-                else
-                {
-                    cnn.Close();
+                    string myConnectionString = "server=7xz.h.filess.io;database=rentacar_wastesugar;uid=rentacar_wastesugar;pwd=d150c35368dc92fa3cc2c09bde449b384fb6b4c3;port=3307;";
+                    MySqlConnection cnn = new MySqlConnection(myConnectionString);
                     cnn.Open();
-                    string hashedPassword = HashPassword(molusturSifre.Text);
-                    MySqlCommand sorgu1 = new MySqlCommand("INSERT INTO users(TC, password, name, surname, tel_no, isBanned) VALUES('" + molusturTC.Text + "', '" + hashedPassword + "', '" + molusturAd.Text + "','" + molusturSoyad.Text + "', '" + molusturTelNo.Text + "', '0')", cnn);
-                    sorgu1.ExecuteNonQuery();
-                    MessageBox.Show("Yeni müşteri oluşturuldu.");
-                    cnn.Close();
+                    MySqlCommand sorgu = new MySqlCommand("SELECT * FROM users WHERE TC= '" + molusturTC.Text + "'", cnn);
+                    MySqlDataReader tara = sorgu.ExecuteReader();
+                    if (tara.Read())
+                    {
+                        MessageBox.Show("Böyle bir müşteri zaten mevcut, şifresini unuttuysanız yan taraftan şifresini değiştirebilirsiniz veya müşteriyi kaldırabilirsiniz.", "UYARI", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        cnn.Close();
+                    }
+                    else
+                    {
+                        cnn.Close();
+                        cnn.Open();
+                        string hashedPassword = HashPassword(molusturSifre.Text);
+                        MySqlCommand sorgu1 = new MySqlCommand("INSERT INTO users(TC, password, name, surname, tel_no, isBanned) VALUES('" + molusturTC.Text + "', '" + hashedPassword + "', '" + molusturAd.Text + "','" + molusturSoyad.Text + "', '" + molusturTelNo.Text + "', '0')", cnn);
+                        sorgu1.ExecuteNonQuery();
+                        MessageBox.Show("Yeni müşteri oluşturuldu.", "Başarılı", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        cnn.Close();
+                    }
+                }
+                else if (result == DialogResult.No)
+                {
                 }
             }
         }
@@ -111,32 +122,39 @@ namespace RentACar2023
             }
             else if (deleteTC.Text.Length != 11)
             {
-                MessageBox.Show("Lütfen TC kimlik numaranızı eksiksiz bir şekilde giriniz.");
+                MessageBox.Show("Lütfen TC kimlik numaranızı eksiksiz bir şekilde giriniz.", "UYARI", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
             else
             {
-                string myConnectionString = "server=7xz.h.filess.io;database=rentacar_wastesugar;uid=rentacar_wastesugar;pwd=d150c35368dc92fa3cc2c09bde449b384fb6b4c3;port=3307;";
-                MySqlConnection cnn = new MySqlConnection(myConnectionString);
-                cnn.Open();
-                MySqlCommand sorgu = new MySqlCommand("SELECT * FROM users WHERE TC= '" + deleteTC.Text + "'", cnn);
-                MySqlDataReader tara = sorgu.ExecuteReader();
-                if (tara.Read())
+                DialogResult result = MessageBox.Show("Müşteriyi silmek istediğinize emin misiniz?", "Müşteri Silme Onayı", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+
+                if (result == DialogResult.Yes)
                 {
-                    cnn.Close();
+                    string myConnectionString = "server=7xz.h.filess.io;database=rentacar_wastesugar;uid=rentacar_wastesugar;pwd=d150c35368dc92fa3cc2c09bde449b384fb6b4c3;port=3307;";
+                    MySqlConnection cnn = new MySqlConnection(myConnectionString);
                     cnn.Open();
-                    MySqlCommand sorgu1 = new MySqlCommand("DELETE FROM users WHERE TC= '" + deleteTC.Text + "'", cnn);
-                    MySqlDataReader tara1 = sorgu1.ExecuteReader();
-                    MessageBox.Show("Müşteri başarı ile silindi.");
-                    deleteTC.Clear();
-                    cnn.Close();
+                    MySqlCommand sorgu = new MySqlCommand("SELECT * FROM users WHERE TC= '" + deleteTC.Text + "'", cnn);
+                    MySqlDataReader tara = sorgu.ExecuteReader();
+                    if (tara.Read())
+                    {
+                        cnn.Close();
+                        cnn.Open();
+                        MySqlCommand sorgu1 = new MySqlCommand("DELETE FROM users WHERE TC= '" + deleteTC.Text + "'", cnn);
+                        MySqlDataReader tara1 = sorgu1.ExecuteReader();
+                        MessageBox.Show("Müşteri başarı ile silindi.", "Başarılı", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        deleteTC.Clear();
+                        cnn.Close();
+                    }
+                    else
+                    {
+                        MessageBox.Show("Böyle bir müşteri yok", "UYARI", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        cnn.Close();
+                    }
                 }
-                else
+                else if (result == DialogResult.No)
                 {
-                    MessageBox.Show("Böyle bir müşteri yok", "asd", MessageBoxButtons.YesNo, MessageBoxIcon.Error);
-                    cnn.Close();
                 }
             }
-        
         }
         private string HashPassword(string password)
         {
@@ -229,6 +247,12 @@ namespace RentACar2023
             }
         }
 
+        private void geriBTN_Click_1(object sender, EventArgs e)
+        {
+            MusteriForm girisSayfa = new MusteriForm();
+            girisSayfa.Show();
+            this.Hide();
+        }
     }
 
 }
